requirements.txt
phidata
openai
python-dotenv
faiss-cpu
packaging
duckduckgo-search
fastapi
uvicorn
groq
langchain
requests
boto3
botocore
pillow
numpy
pandas
transformers
crawl4ai
langchain-community
pinecone
streamlit
deepseek
tavily-python
selenium
markitdown
playwright
beautifulsoup4
faiss-cpu
pandas
transformers 
pillow
baidu
baidusearch
googlesearch-python
pycountry
langchain-community 
langchain-openai
google-search-results
sentence_transformers
IPython
nest_asyncio
asyncpraw